pyproject.toml
[project]
name = "auth-app"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.4,<6.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "dj-database-url (>=3.0.1,<4.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "django-restframework (>=0.0.1,<0.0.2)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
